+++ hopsworks-web/yo/app/scripts/controllers/projectCtrl.js
@@ -377,6 +377,7 @@ angular.module('hopsWorksApp')
             };
 
             self.goToPython = function () {
+              self.toggleKibanaNavBar();
               self.goToUrl('python');
             };
